+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - added Sizers (AllIns, Fixed)
 - added different execution times (at close, high, low, mid)
 - now all orders are executed. removed cancellation in case of reverse in a future bar with an order pending
+- added bmk return to metrics
+- possibility to run multiple backtest via configuration (TOML) file
 
 ## [0.1.4]
 - added volumes to Data